removed unused variables
`byte signalValue` was originally intended to store wall sensor data, but I renamed it to `int wall` to shorten & increase it's capacity so it could hold 2 bytes instead of 1. (may change later to a short)

`int i` was from professor lewis's code which he used to change the pwrLed by values of 16.